% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwedenMap.R
\name{prepare_map_data}
\alias{prepare_map_data}
\title{This codes merges dataset by knkod. Taken from https://github.com/reinholdsson/swemaps and https://github.com/reinholdsson/rkolada}
\usage{
prepare_map_data(x)
}
\description{
This codes merges dataset by knkod. Taken from https://github.com/reinholdsson/swemaps and https://github.com/reinholdsson/rkolada
}
