% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwedenMap.R
\name{SwedenMap}
\alias{SwedenMap}
\title{A function to plot yearly data from www.kolada.se. Most code is taken from https://github.com/reinholdsson/swemaps and https://github.com/reinholdsson/rkolada}
\usage{
SwedenMap(name = NULL, kpiID = NULL, year = 2015, gif = FALSE,
  value = FALSE, znorm = FALSE, outfile = TRUE)
}
\arguments{
\item{name}{Name of outpufile/plot title. Required. Default is NULL.}

\item{kpiID}{kpiIDs from www.kolada.se. Can be found in the KPI_id files. Required. Default is NULL.}

\item{year}{Year to be plotted as interactive and PDF. Default to 2016.}

\item{gif}{Plots longitudinal data of all years included. Takes substanstial time. Default is FALSE.}

\item{value}{Plots values instead of percentage. Default is FALSE.}

\item{znorm}{Plots z-values per group and year. Default is FALSE.}

\item{outfile}{Writes output to files. Chose FALSE for Rmarkdown document. Default is TRUE.}
}
\description{
This function plots yearly data from  www.kolada.se. Output is an interactive html document, and a static pdf.
If gif=TRUE the function also plots a longitudinal gif of the data
}
\examples{
cat_function()
}
\keyword{cats}
